Added absolute paths and refactore all imports
