+++ mobile/babel.config.js
@@ -2,6 +2,32 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: ["babel-preset-expo"],
-    plugins: [["module-resolver", { root: ["."] }]],
+    plugins: [
+      [
+        "module-resolver",
+        {
+          root: ["."],
+          extensions: [
+            ".ios.ts",
+            ".android.ts",
+            ".ts",
+            ".ios.tsx",
+            ".android.tsx",
+            ".tsx",
+            ".jsx",
+            ".js",
+            ".json",
+          ],
+          alias: {
+            "@assets": "./src/assets/",
+            "@screens": "./src/screens",
+            "@navigation": "./src/navigation",
+            "@components": "./src/components",
+            "@store": "./src/store",
+            "@shared": "./src/shared",
+          },
+        },
+      ],
+    ],
   };
 };
