+++ mobile/package.json
@@ -27,12 +27,12 @@
     "typesafe-actions": "^5.1.0"
   },
   "devDependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/redux": "^3.6.0",
     "@babel/core": "^7.9.0",
     "@babel/eslint-parser": "^7.15.0",
+    "@types/axios": "^0.14.0",
     "@types/react": "~16.9.35",
     "@types/react-native": "~0.63.2",
+    "@types/redux": "^3.6.0",
     "babel-plugin-module-resolver": "^3.2.0",
     "eslint": "^7.32.0",
     "eslint-plugin-react-hooks": "^4.2.0"
