+++ mobile/src/components/shared/ProcessIndicator/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Animated, Pressable } from "react-native";
-import { EProcessState } from "../../../shared/EProcessState";
-import { SharedText } from "../Text";
+import { EProcessState } from "@shared/EProcessState";
+import { SharedText } from "@components/shared/Text";
 import { styles } from "./styles";
 import { ProcessColor, ProcessMessage } from "./options";
