+++ mobile/src/components/shared/Text/styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { ColorTheme } from "../../../shared/ColorTheme";
+import { ColorTheme } from "@shared/ColorTheme";
 import { ESharedTextCategory, TSharedTextProp } from "./options";
 
 export const styles = StyleSheet.create({
