+++ mobile/src/screens/Devices.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { DevicesLoader } from "../components/elements/Devices/DevicesLoader";
-import { DevicesList } from "../components/elements/Devices/List";
-import { SearchField } from "../components/elements/Devices/SearchField";
-import { Spacer } from "../components/shared/Spacer";
+import { DevicesLoader } from "@components/elements/Devices/DevicesLoader";
+import { DevicesList } from "@components/elements/Devices/List";
+import { SearchField } from "@components/elements/Devices/SearchField";
+import { Spacer } from "@components/shared/Spacer";
 
 export const DevicesScreen: React.FC = () => (
   <>
