+++ mobile/src/screens/Users.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { UsersList } from "../components/elements/Users/List";
-import { UsersLoader } from "../components/elements/Users/UsersLoader";
-import { Spacer } from "../components/shared/Spacer";
+import { UsersList } from "@components/elements/Users/List";
+import { UsersLoader } from "@components/elements/Users/UsersLoader";
+import { Spacer } from "@components/shared/Spacer";
 
 export const UsersScreen: React.FC = () => (
   <>
