+++ mobile/src/shared/interfaces/IDevice.ts
