+++ mobile/src/shared/interfaces/IUser.ts
