+++ mobile/src/shared/selectors/useDevicesSelector.ts
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import { IState } from "../../store/IState";
+import { IState } from "@store/IState";
 
 const selector = (state: IState) => state.DevicesReducer;
