+++ mobile/src/store/middlewares/axios/index.ts
@@ -4,7 +4,7 @@ import { axiosErrorSuffix, axiosSuccessSuffix } from "./suffixes";
 
 //baseURL has IP-address because it was tested on real device
 const clients = axios.create({
-  baseURL: "http://172.20.10.2:4000/",
+  baseURL: "http://172.20.10.9:4000/",
   responseType: "json",
   timeout: 1000,
   headers: {
