+++ mobile/src/store/modules/Devices/IDevicesState.ts
@@ -1,5 +1,5 @@
-import { EProcessState } from "../../../shared/EProcessState";
-import { IDevice } from "../../../shared/IDevice";
+import { EProcessState } from "@shared/EProcessState";
+import { IDevice } from "@shared/interfaces/IDevice";
 
 export interface IDevicesState {
   devices: IDevice[];
