+++ mobile/src/store/modules/Users/IUsersState.ts
@@ -1,5 +1,5 @@
-import { EProcessState } from "../../../shared/EProcessState";
-import { IUser } from "../../../shared/IUser";
+import { EProcessState } from "@shared/EProcessState";
+import { IUser } from "@shared/interfaces/IUser";
 
 export interface IUsersState {
   users: IUser[];
