+++ mobile/src/types/typesafe-actions.d.ts
@@ -1,7 +1,7 @@
 import { ActionType } from "typesafe-actions";
 
 export type RootAction = ActionType<
-  typeof import("../store/rootAction").default
+  typeof import("@store/rootAction").default
 >;
 
 declare module "typesafe-actions" {
