+++ mobile/tsconfig.json
@@ -1,4 +1,14 @@
 {
-  "compilerOptions": {},
+  "compilerOptions": {
+    "baseUrl": "./src",
+    "paths": {
+      "@assets/*": ["assets/*"],
+      "@components/*": ["components/*"],
+      "@navigation/*": ["navigation/*"],
+      "@screens/*": ["screens/*"],
+      "@shared/*": ["shared/*"],
+      "@store/*": ["store/*"]
+    }
+  },
   "extends": "expo/tsconfig.base"
 }
